Fix a panic in websocket handler
We shouldn't call zncConn.Close() in the error case; this causes a panic
if zncConn is nil. This was happening to me when the grain was being
woken up by a websocket request; ZNC had insufficient time to boot, so
the connection failed.

I've also added some comments re: potential improvements.

Note that I haven't actually tried this; it's late and I need to go to
bed, but I didn't want to forget. But I'm fairly certain it's the right
thing/source of the problem. Now, instead of crashing the grain will
just boot the client, and it should work on a subsequent connection
attempt.