Experimental: add contexts
Rough ideas mostly. Also added the NOTICE. This isn't something I want
to shove on master (not even tested), but I've written it and don't want
to lose it, so it's going on a branch. I am apparently not in the right
state of mind to be hacking on stuff.